requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
aiosqlite
pillow
pypng